+++ core/src/main/java/com/graphhopper/routing/InstructionsFromEdges.java
@@ -24,7 +24,7 @@ import com.graphhopper.storage.NodeAccess;
 import com.graphhopper.util.*;
 import com.graphhopper.util.shapes.GHPoint;
 
-import static com.graphhopper.util.Instruction.EXIT_ROUNDABOUT;
+import static com.graphhopper.util.Instruction.ROUNDABOUT_EXIT;
 import static com.graphhopper.util.Parameters.Details.*;
 
 /**
@@ -40,6 +40,7 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
     private final NodeAccess nodeAccess;
 
     private final InstructionList ways;
+    private final boolean includeRoundaboutExits;
     private final EdgeExplorer outEdgeExplorer;
     private final EdgeExplorer allExplorer;
     private final BooleanEncodedValue roundaboutEnc;
@@ -86,7 +87,7 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
     private static final int MAX_U_TURN_DISTANCE = 35;
 
     public InstructionsFromEdges(Graph graph, Weighting weighting, EncodedValueLookup evLookup,
-                                 InstructionList ways) {
+                                 InstructionList ways, boolean includeRoundaboutExits) {
         this.weighting = weighting;
         this.roundaboutEnc = evLookup.getBooleanEncodedValue(Roundabout.KEY);
         this.roadEnvEnc = evLookup.getEnumEncodedValue(RoadEnvironment.KEY, RoadEnvironment.class);
@@ -96,6 +97,7 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
         this.lanesEnc = evLookup.hasEncodedValue(Lanes.KEY) ? evLookup.getIntEncodedValue(Lanes.KEY) : null;
         this.nodeAccess = graph.getNodeAccess();
         this.ways = ways;
+        this.includeRoundaboutExits = includeRoundaboutExits;
         prevNode = -1;
         prevInRoundabout = false;
         prevName = null;
@@ -106,16 +108,20 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
         allExplorer = graph.createEdgeExplorer();
     }
 
-    /**
-     * @return the list of instructions for this path.
-     */
-    public static InstructionList calcInstructions(Path path, Graph graph, Weighting weighting, EncodedValueLookup evLookup, final Translation tr) {
+    public static InstructionList calcInstructions(Path path, Graph graph, Weighting weighting,
+                                                   EncodedValueLookup evLookup, final Translation tr) {
+        return calcInstructions(path, graph, weighting, evLookup, tr, true);
+    }
+
+    public static InstructionList calcInstructions(Path path, Graph graph, Weighting weighting,
+                                                   EncodedValueLookup evLookup, final Translation tr,
+                                                   boolean includeRoundaboutExits) {
         final InstructionList ways = new InstructionList(tr);
         if (path.isFound()) {
             if (path.getEdgeCount() == 0) {
                 ways.add(new FinishInstruction(graph.getNodeAccess(), path.getEndNode()));
             } else {
-                path.forEveryEdge(new InstructionsFromEdges(graph, weighting, evLookup, ways));
+                path.forEveryEdge(new InstructionsFromEdges(graph, weighting, evLookup, ways, includeRoundaboutExits));
             }
         }
         return ways;
@@ -179,7 +185,7 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
             // remark: names and annotations within roundabout are ignored
             if (!prevInRoundabout) //just entered roundabout
             {
-                RoundaboutInstruction roundaboutInstruction = new RoundaboutInstruction(Instruction.USE_ROUNDABOUT, name,
+                RoundaboutInstruction roundaboutInstruction = new RoundaboutInstruction(Instruction.ROUNDABOUT_USE, name,
                         new PointList(10, nodeAccess.is3D()));
                 prevInstructionPrevOrientation = prevOrientation;
                 if (prevInstruction != null) {
@@ -247,12 +253,16 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
             rInstr.setRadian(deltaInOut).setDirOfRotation(deltaOut);
 
             // we use an exit that is not considered as it is inaccessible by car (#3081)
-            if(rInstr.getExitNumber() == 0) rInstr.increaseExitNumber();
+            if (rInstr.getExitNumber() == 0) rInstr.increaseExitNumber();
 
-            prevInstruction = new RoundaboutInstruction(EXIT_ROUNDABOUT, name, new PointList(10, nodeAccess.is3D()));
-            // do not set radian and dir of rotation as exit should not have to know exit_number, turn_angle or dirOfRotation.
+            if (includeRoundaboutExits) {
+                prevInstruction = new RoundaboutInstruction(ROUNDABOUT_EXIT, name, new PointList(10, nodeAccess.is3D()));
+                // do not set radian and dir of rotation as exit should not have to know exit_number, turn_angle or dirOfRotation.
 
-            ways.add(prevInstruction);
+                ways.add(prevInstruction);
+            }
+
+            ((RoundaboutInstruction) prevInstruction).setExited();
             prevInstructionPrevOrientation = prevOrientation;
             prevOrientation = orientation;
             prevInstructionName = prevName;